pyproject.toml
[project]
name = "genie-codex-demo"
version = "0.1.0"
description = "Databricks demo workspace for Genie, Codex, ai-dev-kit, jobs, and pipelines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.102.0",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
